divided form specification (fields) from it's instance (JS/JVM), a bit refactored i18n

+++ core/shared/src/main/scala/pl/pholda/malpompaaligxilo/i18n/I18nable.scala
@@ -1,9 +1,7 @@
 package pl.pholda.malpompaaligxilo.i18n
 
-import pl.pholda.malpompaaligxilo.util._
-
 trait I18nable[T] {
   def apply(implicit lang: Lang): T
 
-  def apply(n: Int)(implicit lang: Lang): T
+  def apply(n: Long)(implicit lang: Lang): T
 }
