+++ core/shared/src/main/scala/pl/pholda/malpompaaligxilo/i18n/NoI18nString.scala
@@ -1,9 +1,7 @@
 package pl.pholda.malpompaaligxilo.i18n
 
-import pl.pholda.malpompaaligxilo.util._
-
 case class NoI18nString(str: String) extends I18nableString {
   override def apply(implicit lang: Lang): String = str
 
-  override def apply(n: Int)(implicit lang: Lang): String = str
+  override def apply(n: Long)(implicit lang: Lang): String = str
 }
